Move harvester_farmer_environment to conftest.py.
