+++ tests/conftest.py
@@ -20,9 +20,15 @@ from chia.clvm.spend_sim import CostLogger
 
 # Set spawn after stdlib imports, but before other imports
 from chia.consensus.constants import ConsensusConstants
+from chia.farmer.farmer import Farmer
+from chia.farmer.farmer_api import FarmerAPI
 from chia.full_node.full_node import FullNode
 from chia.full_node.full_node_api import FullNodeAPI
+from chia.harvester.harvester import Harvester
+from chia.harvester.harvester_api import HarvesterAPI
 from chia.protocols import full_node_protocol
+from chia.rpc.farmer_rpc_client import FarmerRpcClient
+from chia.rpc.harvester_rpc_client import HarvesterRpcClient
 from chia.rpc.wallet_rpc_client import WalletRpcClient
 from chia.seeder.crawler import Crawler
 from chia.seeder.crawler_api import CrawlerAPI
@@ -890,3 +896,38 @@ async def simulation(consensus_mode, bt):
         pytest.skip("Skipping this run. This test only supports one running at a time.")
     async for _ in setup_full_system(test_constants_modified, bt, db_version=1):
         yield _
+
+
+HarvesterFarmerEnvironment = Tuple[
+    Service[Farmer, FarmerAPI], FarmerRpcClient, Service[Harvester, HarvesterAPI], HarvesterRpcClient, BlockTools
+]
+
+
+@pytest_asyncio.fixture(scope="function")
+async def harvester_farmer_environment(
+    farmer_one_harvester: Tuple[List[Service[Harvester, HarvesterAPI]], Service[Farmer, FarmerAPI], BlockTools],
+    self_hostname: str,
+) -> AsyncIterator[HarvesterFarmerEnvironment]:
+    harvesters, farmer_service, bt = farmer_one_harvester
+    harvester_service = harvesters[0]
+
+    assert farmer_service.rpc_server is not None
+    farmer_rpc_cl = await FarmerRpcClient.create(
+        self_hostname, farmer_service.rpc_server.listen_port, farmer_service.root_path, farmer_service.config
+    )
+    assert harvester_service.rpc_server is not None
+    harvester_rpc_cl = await HarvesterRpcClient.create(
+        self_hostname, harvester_service.rpc_server.listen_port, harvester_service.root_path, harvester_service.config
+    )
+
+    async def have_connections() -> bool:
+        return len(await farmer_rpc_cl.get_connections()) > 0
+
+    await time_out_assert(15, have_connections, True)
+
+    yield farmer_service, farmer_rpc_cl, harvester_service, harvester_rpc_cl, bt
+
+    farmer_rpc_cl.close()
+    harvester_rpc_cl.close()
+    await farmer_rpc_cl.await_closed()
+    await harvester_rpc_cl.await_closed()
